fix: fixed multiple execution of validation function. Should consider more self evident appoach
